WS-034: Refactor admin buttons
